set int dtype for time features

+++ declafe/feature_gen/unary/times/DayOfMonthFeature.py
@@ -12,7 +12,7 @@ class DayOfMonthFeature(UnaryFeature):
   def gen_unary(self, ser: np.ndarray) -> np.ndarray:
     gen = np.frompyfunc(
         lambda x: datetime.utcfromtimestamp(x / 1000_000_000).day, 1, 1)
-    return gen(ser)
+    return gen(ser).astype(np.int64)
 
   @property
   def name(self) -> str:
