+++ declafe/feature_gen/unary/times/HourFeature.py
@@ -12,7 +12,7 @@ class HourFeature(UnaryFeature):
   def gen_unary(self, ser: np.ndarray) -> np.ndarray:
     gen = np.frompyfunc(
         lambda x: datetime.utcfromtimestamp(x / 1000_000_000).hour, 1, 1)
-    return gen(ser)
+    return gen(ser).astype(np.int64)
 
   @property
   def name(self) -> str:
