+++ declafe/feature_gen/unary/times/WeekOfYear.py
@@ -18,9 +18,10 @@ class WeekOfYearFeature(UnaryFeature):
 
   def gen_unary(self, ser: np.ndarray) -> np.ndarray:
     gen = np.frompyfunc(
-        lambda x: datetime.utcfromtimestamp(x / 1000_000_000).isocalendar()[1], 1, 1)
+        lambda x: datetime.utcfromtimestamp(x / 1000_000_000).isocalendar()[1],
+        1, 1)
 
-    return gen(ser)
+    return gen(ser).astype(np.int64)
 
   @property
   def name(self) -> str:
